+++ lib/api_test_helper/tests/not_empty.rb
@@ -37,6 +37,8 @@ module ApiTestHelper
 
         @failed = value.empty?
         success?
+      rescue TypeError
+        @failed = true
       end
 
       def success?