+++ lib/api_test_helper/tests/regexp.rb
@@ -42,6 +42,8 @@ module ApiTestHelper
           @failed = true
         end
         success?
+      rescue TypeError
+        @failed = true
       end
 
       def success?